Fix unable to edit thread replies.
